refactor cli . inf server tests
